move struct MoveKind to mod chess::move
