+++ chess/src/move.rs
@@ -1,10 +1,52 @@
 use crate::{Promotion, Role, Square};
+use bitflags::bitflags;
 use derive_more::{DebugCustom, Deref, Display, Error};
 use shakmaty as sm;
 use std::str::FromStr;
 use test_strategy::Arbitrary;
 use util::{Binary, Bits};
 use vampirc_uci::UciMove;
+
+bitflags! {
+    /// Characteristics of a [`Move`] in the context of a [`Position`][`crate::Position`].
+    #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
+    pub struct MoveKind: u8 {
+        const ANY =         0b00000001;
+        const CASTLE =      0b00000010;
+        const PROMOTION =   0b00000100;
+        const CAPTURE =     0b00001000;
+    }
+}
+
+#[doc(hidden)]
+impl From<&sm::Move> for MoveKind {
+    #[inline]
+    fn from(m: &sm::Move) -> Self {
+        let mut kind = Self::ANY;
+
+        if m.is_castle() {
+            kind |= MoveKind::CASTLE
+        }
+
+        if m.is_promotion() {
+            kind |= MoveKind::PROMOTION
+        }
+
+        if m.is_capture() {
+            kind |= MoveKind::CAPTURE;
+        }
+
+        kind
+    }
+}
+
+#[doc(hidden)]
+impl From<&mut sm::Move> for MoveKind {
+    #[inline]
+    fn from(m: &mut sm::Move) -> Self {
+        (&*m).into()
+    }
+}
 
 /// A chess move.
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash, Arbitrary, Deref)]
